refactor(recovery): move the reusable parts of subnet splitting into a crate of their own
The subnet splitting tool carries a number of helpers that have nothing to do
with splitting a subnet: reading the state tree the NNS signed and the subnet
public key in it, validating the CUP a subnet halted at against that key and
the state it halted at against that CUP, running the `state_tool` commands on a
checkpoint, reading a CUP and the state hash of a checkpoint, proposing to halt
a subnet at its next CUP height, printing a dashboard URL and waiting for the
operator to confirm it, and reading a part of the registry back after a
proposal.

They move to `rs/recovery/subnet_tools`, which `subnet_splitting` now depends
on. Any other tool that reshapes the canister id ranges of a subnet -- halting
it at a CUP, working on the state it halted at, and validating what it produced
-- can then use them instead of copying them.

No behavior change: the code moved as it was, except that the items it exposes
became `pub` and that the log and error messages of the CUP validation no
longer say "split" and "source subnet", as they now describe a subnet that was
halted at a CUP, whichever operation halted it. `state_tool_lib` is visible to
`//rs/recovery:__subpackages__` rather than to `subnet_splitting` by name.
