+++ rs/recovery/subnet_splitting/BUILD.bazel
@@ -13,11 +13,10 @@ rust_library(
     visibility = ["//rs:system-tests-pkg"],
     deps = [
         # Keep sorted.
-        "//rs/crypto/utils/threshold_sig",
-        "//rs/crypto/utils/threshold_sig_der",
         "//rs/monitoring/metrics",
         "//rs/protobuf",
         "//rs/recovery",
+        "//rs/recovery/subnet_tools",
         "//rs/registry/routing_table",
         "//rs/registry/subnet_type",
         "//rs/state_layout",
@@ -29,10 +28,7 @@ rust_library(
         "@crate_index//:anyhow",
         "@crate_index//:clap",
         "@crate_index//:csv",
-        "@crate_index//:hex",
-        "@crate_index//:ic-agent",
         "@crate_index//:serde",
-        "@crate_index//:serde_cbor",
         "@crate_index//:slog",
         "@crate_index//:strum",
         "@crate_index//:url",
@@ -47,11 +43,10 @@ rust_binary(
         # Keep sorted.
         ":subnet_splitting",
         "//rs/canister_sandbox:backend_lib",
-        "//rs/crypto/utils/threshold_sig",
-        "//rs/crypto/utils/threshold_sig_der",
         "//rs/monitoring/metrics",
         "//rs/protobuf",
         "//rs/recovery",
+        "//rs/recovery/subnet_tools",
         "//rs/registry/routing_table",
         "//rs/registry/subnet_type",
         "//rs/state_layout",
@@ -63,10 +58,7 @@ rust_binary(
         "@crate_index//:anyhow",
         "@crate_index//:clap",
         "@crate_index//:csv",
-        "@crate_index//:hex",
-        "@crate_index//:ic-agent",
         "@crate_index//:serde",
-        "@crate_index//:serde_cbor",
         "@crate_index//:slog",
         "@crate_index//:strum",
         "@crate_index//:url",
@@ -79,11 +71,10 @@ rust_test(
     crate = "subnet_splitting",
     deps = [
         # Keep sorted.
-        "//rs/crypto/utils/threshold_sig",
-        "//rs/crypto/utils/threshold_sig_der",
         "//rs/monitoring/metrics",
         "//rs/protobuf",
         "//rs/recovery",
+        "//rs/recovery/subnet_tools",
         "//rs/registry/routing_table",
         "//rs/registry/subnet_type",
         "//rs/state_layout",
@@ -96,10 +87,7 @@ rust_test(
         "@crate_index//:anyhow",
         "@crate_index//:clap",
         "@crate_index//:csv",
-        "@crate_index//:hex",
-        "@crate_index//:ic-agent",
         "@crate_index//:serde",
-        "@crate_index//:serde_cbor",
         "@crate_index//:slog",
         "@crate_index//:strum",
         "@crate_index//:url",
@@ -126,11 +114,10 @@ rust_ic_test(
     deps = [
         # Keep sorted.
         ":subnet_splitting",
-        "//rs/crypto/utils/threshold_sig",
-        "//rs/crypto/utils/threshold_sig_der",
         "//rs/monitoring/metrics",
         "//rs/protobuf",
         "//rs/recovery",
+        "//rs/recovery/subnet_tools",
         "//rs/registry/routing_table",
         "//rs/registry/subnet_type",
         "//rs/rust_canisters/proxy_canister:lib",
@@ -150,10 +137,7 @@ rust_ic_test(
         "@crate_index//:candid",
         "@crate_index//:clap",
         "@crate_index//:csv",
-        "@crate_index//:hex",
-        "@crate_index//:ic-agent",
         "@crate_index//:serde",
-        "@crate_index//:serde_cbor",
         "@crate_index//:slog",
         "@crate_index//:strum",
         "@crate_index//:url",
