+++ rs/recovery/subnet_splitting/Cargo.toml
@@ -10,11 +10,7 @@ documentation.workspace = true
 anyhow = { workspace = true }
 clap = { workspace = true }
 csv = { workspace = true }
-hex = { workspace = true }
-ic-agent = { workspace = true }
 ic-base-types = { path = "../../types/base_types/" }
-ic-crypto-utils-threshold-sig = { path = "../../crypto/utils/threshold_sig" }
-ic-crypto-utils-threshold-sig-der = { path = "../../crypto/utils/threshold_sig_der" }
 ic-metrics = { path = "../../monitoring/metrics" }
 ic-protobuf = { path = "../../protobuf" }
 ic-recovery = { path = "../" }
@@ -23,10 +19,10 @@ ic-registry-subnet-type = { path = "../../registry/subnet_type" }
 ic-state-layout = { path = "../../state_layout" }
 ic-state-manager = { path = "../../state_manager" }
 ic-state-tool = { path = "../../state_tool" }
+ic-subnet-tools = { path = "../subnet_tools" }
 ic-types-cycles = { path = "../../types/cycles" }
 ic-types = { path = "../../types/types" }
 serde = { workspace = true }
-serde_cbor = { workspace = true }
 slog = { workspace = true }
 strum = { workspace = true }
 strum_macros = { workspace = true }
