+++ rs/recovery/subnet_splitting/src/lib.rs
@@ -1,11 +1,8 @@
 pub mod post_split_estimations;
 pub mod subnet_splitting;
 pub mod utils;
-pub mod validation;
 
 mod admin_helper;
-mod agent_helper;
 mod layout;
-mod state_tool_helper;
 mod steps;
 mod target_subnet;
