+++ rs/recovery/subnet_splitting/src/utils.rs
@@ -1,31 +1,12 @@
 use ic_base_types::SubnetId;
-use ic_protobuf::types::v1 as pb;
 use ic_recovery::{
     error::{RecoveryError, RecoveryResult},
     file_sync_helper::read_file,
     util::subnet_id_from_str,
 };
 use ic_registry_routing_table::CanisterIdRange;
-use ic_state_manager::manifest::{manifest_from_path, manifest_hash};
-use ic_types::{Time, consensus::CatchUpPackage};
 
-use std::{fmt::Display, path::Path};
-
-pub(crate) fn get_batch_time_from_cup(cup_path: &Path) -> RecoveryResult<Time> {
-    get_cup(cup_path).map(|cup| cup.content.block.as_ref().context.time)
-}
-
-pub(crate) fn get_cup(cup_path: &Path) -> RecoveryResult<CatchUpPackage> {
-    let cup_proto = pb::CatchUpPackage::read_from_file(cup_path)
-        .map_err(|err| cup_error("Failed to decode the CUP file", cup_path, err))?;
-
-    CatchUpPackage::try_from(&cup_proto)
-        .map_err(|err| cup_error("Failed to deserialize the CUP file", cup_path, err))
-}
-
-fn cup_error(message: impl Display, cup_path: &Path, error: impl Display) -> RecoveryError {
-    RecoveryError::UnexpectedError(format!("{} ({}): {}", message, cup_path.display(), error))
-}
+use std::path::Path;
 
 pub(crate) fn canister_id_range_to_string(canister_id_range: &CanisterIdRange) -> String {
     format!("{}:{}", canister_id_range.start, canister_id_range.end)
@@ -36,21 +17,6 @@ pub fn canister_id_ranges_to_strings(canister_id_ranges: &[CanisterIdRange]) -> 
         .iter()
         .map(canister_id_range_to_string)
         .collect::<Vec<_>>()
-}
-
-/// Computes the state hash of the given checkpoint.
-pub(crate) fn get_state_hash(checkpoint_dir: impl AsRef<Path>) -> RecoveryResult<String> {
-    let manifest = manifest_from_path(checkpoint_dir.as_ref()).map_err(|e| {
-        RecoveryError::CheckpointError(
-            format!(
-                "Failed to read the manifest from path {}",
-                checkpoint_dir.as_ref().display()
-            ),
-            e,
-        )
-    })?;
-
-    Ok(hex::encode(manifest_hash(&manifest)))
 }
 
 /// Parses the output of `state-tool split-manifests` and finds the expected root hash of the split
