+++ rs/recovery/subnet_tools/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "ic-subnet-tools"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+description.workspace = true
+documentation.workspace = true
+
+[dependencies]
+hex = { workspace = true }
+ic-agent = { workspace = true }
+ic-base-types = { path = "../../types/base_types/" }
+ic-crypto-utils-threshold-sig = { path = "../../crypto/utils/threshold_sig" }
+ic-crypto-utils-threshold-sig-der = { path = "../../crypto/utils/threshold_sig_der" }
+ic-protobuf = { path = "../../protobuf" }
+ic-recovery = { path = "../" }
+ic-registry-routing-table = { path = "../../registry/routing_table" }
+ic-registry-subnet-type = { path = "../../registry/subnet_type" }
+ic-state-manager = { path = "../../state_manager" }
+ic-state-tool = { path = "../../state_tool" }
+ic-types = { path = "../../types/types" }
+serde_cbor = { workspace = true }
+slog = { workspace = true }
+url = { workspace = true }
